+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+10.11.2016: 0.9.1 release
+    - [break] composer.lock file was removed
+    - [brake] `src\helper\PhpVersionChecker.php` was removed
+    - [brake] `src\helper\PhpVersionCheckerInterface.php` was removed
+    - [brake] `devgroup\grayii\GelfTarget:isThrowableMessage` method was removed 
+
 10.11.2016: 0.9.0 release
     - php 5.6 supports
     - php 7.0 supports